Cache AHPs Spotijack session to reduce CPU overhead when polling
Profiling with instruments revealed that the pollAudioHijackPro method accounted
for 64% of the total applicationPollingTimerFired method's call time. Almost all
of this time was spent by the getFirstSpotijackSession method calling method on
SBArray. This could be a performance regression in Mojave as I don't remember
SBArray being this slow (I don't have anything to test with though).

This commit caches the Spotijack session returned on the first call to
getFirstSpotijackSession (now renamed to a property _spotijackSession). As the
cached SBObject refers to a session by index in Audio Hijack Pro, the cached
object is invalidated if its name changes from "Spotijack". This would indicate
that the Spotijack session has been moved or deleted. This commit reduces the
time spent in pollAudioHijackPro in the applicationPollingTimerFired method to
36% of the total call time.