Significantly reduce CPU overhead when polling Spotify for track changes
This is a big performance win. Starting from Mojave (maybe? Never noticed it
before), accessing properties on SBObject is expensive. This commit optimises
property access in Spotijack's pollSpotify method to minimise CPU overhead.

Before this commit, pollSpotify accounted for 64% of the time spent in the
`applicationPollingTimerFired` method. With this commit, pollSpotify accounts
for 18% of the time spent in `applicationPollingTimerFired`.

Most of the performance gain comes from avoiding creating a `StaticSpotifyTrack`
whenever possible. This object is very expensive to construct because it
accesses 12 (!) properties on an SBObject when initialised. With this commit,
the Spotify polling method only constructs a `StaticSpotifyTrack` if the track
ID changes. It doesn't check any other properties.