Stop checking name of cached AHP session
In my ongoing series of performance operations to the polling loop, it seems
that just checking the name of the cached session is slow since it involves an
RPC. Removing the name check reduces the time it takes to poll AHP by
50%. Importantly, on my system this brings the total polling time down to <10 ms
so it keeps up with the polling rate.

As a result of this change, removing or re-ordering the Spotijack session in AHP
will break Spotijack but a simple restart will fix it. I think this is a
worthwhile price for reduced polling overhead.