Fix crash when manually changing tracks in Spotify
Ack, this undoes some of the performance gains in db234af. Whenever you manually
change a track in Spotify (i.e., double click on a track to play it),
`SpotifyTrack.get()` returns `nil`.

It looks like when you manually change track, Spotify pauses, sets the track to
`nil`, selects the new track and then un-pauses. For some reason it doesn't give
a resolvable `SpotifyTrack` when that happens.