Flesh out recording saving system
Now, recordings are saved to a temporary directory using a GUID as a
filename. This should pretty much eliminate file naming conflicts. When the
recording finishes, the recording session renames the recording and places it in
the appropriate output directory.

Recordings can now be placed in arbitrarily nested directories. The session will
create any intermediate directories before moving the recordings. If a recording
with the same name already exists in the target directory, the session will
append a number to the file to resolve any conflicts.